clean up collabPage in frontend
