+++ frontend/public/runtime-env.js
@@ -1,6 +1,1 @@
-window.__RUNTIME_CONFIG__ = {
-  REACT_APP_URI_USER_SVC: '"http://localhost:8000"',
-  REACT_APP_URI_MATCH_SVC: '"http://localhost:8001"',
-  REACT_APP_URI_QUESTION_SVC: '"http://localhost:8002"',
-  REACT_APP_URI_COLLABORATION_SVC: '"http://localhost:8005"',
-};
+window.__RUNTIME_CONFIG__ = {"REACT_APP_URI_USER_SVC":"http://localhost:8000","REACT_APP_URI_MATCH_SVC":"http://localhost:8001"};
