+++ frontend/src/App.js
@@ -6,9 +6,8 @@ import PublicRoutes from "./routes/PublicRoutes";
 import LoginPage from "./views/login/LoginPage";
 import SignupPage from "./views/signup/SignupPage";
 import LandingPage from "./views/landing/LandingPage";
-import CollabPage from "./views/collab/CollabPage";
 import MatchingPage from "./views/MatchingPage";
-import CollabPage2 from "./views/collab/CollabPage2";
+import CollabPage from "./views/collab/CollabPage";
 import HistoryPage from "./views/history/HistoryPage";
 import ForgotPasswordPage from "./views/forgotPassword/ForgotPasswordPage";
 import "./App.scss";
@@ -19,10 +18,9 @@ function App() {
       <Routes>
         <Route element={<PrivateRoutes />}>
           <Route path="/match" element={<MatchingPage />} />
-          <Route path="/collab" element={<CollabPage />} />
           <Route path="/history" element={<HistoryPage />} />
           <Route element={<CollabRoute />}>
-            <Route path="/collab2" element={<CollabPage2 />} />
+            <Route path="/collab" element={<CollabPage />} />
           </Route>
         </Route>
         <Route element={<PublicRoutes />}>
