+++ frontend/src/views/collab/CollabPage.js
@@ -1,78 +1,56 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import Button from "react-bootstrap/esm/Button";
-import CodeEditorWindow from "../../components/CodeEditorWindow";
-import LeaveRoomModal from "../../components/LeaveRoomModal";
 import NavBar from "../../components/NavBar";
 import QuestionCard from "../../components/QuestionCard";
 import ChatWindow from "../../components/ChatWindow";
-
-import { languageOptions } from "../../utils/codeEditorUtils/languageOptions";
-import { defineTheme } from "../../utils/codeEditorUtils/defineTheme";
+import CollabEditor from "../../components/CollabEditor";
+import LeaveRoomModal from "../../components/LeaveRoomModal";
+import NotifyUserLeftModal from "../../components/NotifyUserLeftModal";
+import { socketSelector } from "../../stores/socket/socket.slice";
+import { useSelector } from "react-redux";
+import { matchSelector } from "../../stores/match/match.slice";
+import axios from "axios";
+import configs from "../../utils/configs";
+const config = configs[process.env.NODE_ENV];
 
 import "./CollabPage.scss";
 
-const javascriptDefault = `/**
-* Language: Javascript
-*/
-
-console.log('Hello World!');
-`;
-
-function CollabPage() {
+function CollabPage2() {
   const [showLeaveRoomModal, setShowLeaveRoomModal] = useState(false);
-  const [code, setCode] = useState(javascriptDefault);
-  const [theme, setTheme] = useState("cobalt");
-  const [language, setLanguage] = useState(languageOptions[0]);
+  const [showUserLeftModal, setShowUserLeftModal] = useState(false);
 
-  const onChange = (action, data) => {
-    switch (action) {
-      case "code": {
-        setCode(data);
-        break;
-      }
-      default: {
-        console.warn("case not handled!", action, data);
-      }
-    }
-  };
+  const { socket } = useSelector(socketSelector);
+  const { matchId } = useSelector(matchSelector);
 
-  const handleLanguageChange = (selectedLanguage) => {
-    setLanguage(selectedLanguage);
-  };
+  socket.on("other user left room", () => {
+    setShowUserLeftModal(true);
+  });
 
-  const handleThemeChange = (selectedTheme) => {
-    if (["light", "vs-dark"].includes(selectedTheme.value)) {
-      setTheme(selectedTheme);
-    } else {
-      defineTheme(selectedTheme.value).then(() => setTheme(selectedTheme));
-    }
-  };
-
-  useEffect(() => {
-    defineTheme("oceanic-next").then(() =>
-      setTheme({ value: "oceanic-next", label: "Oceanic Next" }),
-    );
-  }, []);
+  socket.on("last user left room", () => {
+    axios.delete(config.QUESTION_SVC_BASE_URL + "/question-api/assigned-question", {
+      data: {
+        matchId: matchId,
+      },
+    });
+  });
 
   return (
     <>
-      <div className="Collab-container">
-        <NavBar />
-        <QuestionCard />
-        <div className="Editor-chat-container">
-          <CodeEditorWindow
-            code={code}
-            onChange={onChange}
-            language={language}
-            theme={theme}
-            handleLanguageChange={handleLanguageChange}
-            handleThemeChange={handleThemeChange}
-          />
-          <ChatWindow />
+      <div className="Collab2-container">
+        <NavBar isCollabPage />
+        <div className="Collab2-content-div">
+          <div className="Collab2-left-div">
+            <QuestionCard containerId="Collab2-qn-card-container" />
+            <ChatWindow />
+          </div>
+
+          <div className="Collab2-right-div">
+            <CollabEditor />
+          </div>
         </div>
-        <div className="Collab-actions-container p-3">
+        <div className="Collab2-footer-div px-3 py-2">
           <Button
-            className="btn-leave-room"
+            className="btn-leave-room btn-sm"
             variant="outline-danger"
             onClick={() => setShowLeaveRoomModal(true)}
           >
@@ -80,9 +58,14 @@ function CollabPage() {
           </Button>
         </div>
       </div>
-      <LeaveRoomModal handleClose={() => setShowLeaveRoomModal(false)} show={showLeaveRoomModal} />
+
+      <LeaveRoomModal show={showLeaveRoomModal} handleClose={() => setShowLeaveRoomModal(false)} />
+      <NotifyUserLeftModal
+        handleClose={() => setShowUserLeftModal(false)}
+        show={showUserLeftModal}
+      />
     </>
   );
 }
 
-export default CollabPage;
+export default CollabPage2;
